ignore nexthop attribute when NLRI is present
According to section 1.3 of RFC2858, an UPDATE message that carries no NLRI,
other than the one encoded in the MP_REACH_NLRI attribute, should not carry
the NEXT_HOP attribute. If such a message contains the NEXT_HOP attribute,
the BGP speaker that receives the message should ignore this attribute.